requirements.txt
pyboy>=2.6.0,<3
mistralai>=1.9.0,<2
gymnasium>=0.29.1,<1.1
numpy>=1.26.0,<3
stable-baselines3>=2.3.2,<3
pillow>=10,<12
